prompts/intent_prompts.py
# prompts/intent_prompts.py
from langchain_core.prompts import ChatPromptTemplate



INTENT_CLASSIFICATION_SYSTEM = """
You are a senior technical triage lead. Your job is to classify developer queries into one of the following categories:

- BUG: Reports of unexpected behavior or errors in code.
- ARCHITECTURE: High-level design, system integration, or scalability questions.
- CONCEPT: General explanations of a technology, language, or framework.
- COMPARISON: Benchmarking or choosing between two or more tools/stacks.
- RESEARCH: Deep dives into emerging tech, RFCs, or complex documentation analysis.

If the query is too vague to classify, return 'unclear'.
"""

INTENT_PROMPT = ChatPromptTemplate.from_messages([
    ("system", INTENT_CLASSIFICATION_SYSTEM),
    ("user", "Classify this query: {query}")
])