requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.30
groq>=0.1.0
qdrant-client==1.17.0
fastembed>=0.2.0
tavily-python>=0.3.0
duckduckgo-search>=3.9.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
# transformers removed - using Groq API instead of HuggingFace